organize settings and home modules
