+++ desktop/src/pages/settings/sections/models.tsx
@@ -1,0 +1,125 @@
+import { useTranslation } from 'react-i18next'
+import { ReactComponent as FolderIcon } from '~/icons/folder.svg'
+import { ReactComponent as LinkIcon } from '~/icons/link.svg'
+import { ReactComponent as WrenchIcon } from '~/icons/wrench.svg'
+import { Button } from '~/components/ui/button'
+import { Input } from '~/components/ui/input'
+import { Label } from '~/components/ui/label'
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '~/components/ui/select'
+import { SectionCard, type SettingsViewModel } from './shared'
+
+export function ModelsSection({ vm }: { vm: SettingsViewModel }) {
+	const { t } = useTranslation()
+	return (
+<div className="space-y-5">
+							<SectionCard>
+								<div className="space-y-5">
+									<div className="space-y-2">
+										<Label>{t('common.download-model')}</Label>
+										<div className="flex items-center gap-2">
+											<Input
+												type="text"
+												value={vm.downloadURL}
+												onChange={(event) => vm.setDownloadURL(event.target.value)}
+												placeholder={t('common.paste-model-link')}
+												onKeyDown={(event) => (event.key === 'Enter' ? vm.downloadModel() : null)}
+											/>
+											<Button variant="default" size="icon" onClick={vm.downloadModel} className="shrink-0">
+												<svg
+													aria-hidden="true"
+													focusable="false"
+													role="img"
+													className="octicon octicon-download"
+													viewBox="0 0 16 16"
+													width="16"
+													height="16"
+													fill="currentColor">
+													<path d="M2.75 14A1.75 1.75 0 0 1 1 12.25v-2.5a.75.75 0 0 1 1.5 0v2.5c0 .138.112.25.25.25h10.5a.25.25 0 0 0 .25-.25v-2.5a.75.75 0 0 1 1.5 0v2.5A1.75 1.75 0 0 1 13.25 14Z"></path>
+													<path d="M7.25 7.689V2a.75.75 0 0 1 1.5 0v5.689l1.97-1.969a.749.749 0 1 1 1.06 1.06l-3.25 3.25a.749.749 0 0 1-1.06 0L4.22 6.78a.749.749 0 1 1 1.06-1.06l1.97 1.969Z"></path>
+												</svg>
+											</Button>
+										</div>
+									</div>
+
+									<div className="space-y-2">
+										<Label>{t('common.select-model')}</Label>
+										<Select
+											value={vm.preference.modelPath ?? undefined}
+											onValueChange={(value) => vm.preference.setModelPath(value)}
+											onOpenChange={(open) => {
+												if (open) vm.loadModels()
+											}}>
+											<SelectTrigger>
+												<SelectValue placeholder={t('common.select-model')} />
+											</SelectTrigger>
+											<SelectContent>
+												{vm.models.map((model, index) => (
+													<SelectItem key={index} value={model.path}>
+														{model.name}
+													</SelectItem>
+												))}
+											</SelectContent>
+										</Select>
+									</div>
+
+									{!vm.isMacOS && (
+										<div className="space-y-2">
+											<Label>{t('common.gpu-device')}</Label>
+											{vm.gpuDevices.length > 0 ? (
+												<Select
+													value={vm.preference.gpuDevice != null ? String(vm.preference.gpuDevice) : 'auto'}
+													onValueChange={(value) => {
+														vm.preference.setGpuDevice(value === 'auto' ? null : parseInt(value, 10))
+													}}>
+													<SelectTrigger>
+														<SelectValue placeholder={t('common.gpu-device')} />
+													</SelectTrigger>
+													<SelectContent>
+														<SelectItem value="auto">Auto</SelectItem>
+														{vm.gpuDevices.map((device) => (
+															<SelectItem key={device.index} value={String(device.index)}>
+																{device.description}
+															</SelectItem>
+														))}
+													</SelectContent>
+												</Select>
+											) : (
+												<Input
+													type="number"
+													value={vm.preference.gpuDevice ?? ''}
+													onChange={(e) => {
+														const val = e.target.value
+														vm.preference.setGpuDevice(val === '' ? null : parseInt(val, 10))
+													}}
+													placeholder={t('common.gpu-device-placeholder')}
+												/>
+											)}
+										</div>
+									)}
+
+									<div className="space-y-1 pt-1">
+										<Button
+											variant="ghost"
+											onMouseDown={vm.openModelsUrl}
+											className="h-11 w-full justify-between rounded-lg px-3 font-medium hover:bg-accent/60">
+											{t('common.download-models-link')} <LinkIcon className="h-4 w-4 text-muted-foreground" />
+										</Button>
+										<Button
+											variant="ghost"
+											onMouseDown={vm.openModelPath}
+											className="h-11 w-full justify-between rounded-lg px-3 font-medium hover:bg-accent/60">
+											{t('common.models-folder')} <FolderIcon className="h-4 w-4 text-muted-foreground" />
+										</Button>
+										<Button
+											variant="ghost"
+											onMouseDown={vm.changeModelsFolder}
+											className="h-11 w-full justify-between rounded-lg px-3 font-medium hover:bg-accent/60">
+											{t('common.change-models-folder')} <WrenchIcon className="h-4 w-4 text-muted-foreground" />
+										</Button>
+									</div>
+								</div>
+							</SectionCard>
+
+						</div>
+	)
+}
